Extract client method for segments in SegmentClient class
