+++ tests/clients/BarracksClient.spec.js
@@ -11,13 +11,6 @@ const proxyquire = require('proxyquire').noCallThru();
 chai.use(chaiAsPromised);
 chai.use(sinonChai);
 
-function buildSegment(segmentId) {
-  return {
-    id: segmentId,
-    name: 'Plop'
-  };
-}
-
 describe('Barracks', () => {
 
   let barracks;
@@ -50,6 +43,19 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('setGoogleAnalyticsTrackingId').and.to.be.a('function');
     });
 
+    it('should initialize methods from segmentClient when constructor called', () => {
+      // When
+      const barracksClient = new Barracks();
+
+      // Then
+      expect(barracksClient).to.have.property('createSegment').and.to.be.a('function');
+      expect(barracksClient).to.have.property('editSegment').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getSegmentByName').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getSegment').and.to.be.a('function');
+      expect(barracksClient).to.have.property('getSegments').and.to.be.a('function');
+      expect(barracksClient).to.have.property('setActiveSegments').and.to.be.a('function');
+    });
+
     it('should initialize methods from updateClient when constructor called', () => {
       // When
       const barracksClient = new Barracks();
@@ -63,144 +69,6 @@ describe('Barracks', () => {
       expect(barracksClient).to.have.property('publishUpdate').and.to.be.a('function');
       expect(barracksClient).to.have.property('archiveUpdate').and.to.be.a('function');
       expect(barracksClient).to.have.property('scheduleUpdate').and.to.be.a('function');
-    });
-  });
-
-  describe('#editSegment()', () => {
-
-    const originalSegment = {
-      id: '12345678',
-      name: 'test segment',
-      query: { eq: { unitId: '123456' } }
-    };
-
-    it('should return an error message when edit request fails', done => {
-      // Given
-      const changes = { id: originalSegment.id, name: 'New name' };
-      const error = { message: 'Error !' };
-      barracks.getSegment = sinon.stub().returns(Promise.resolve(originalSegment));
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.editSegment(token, changes).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        try {
-          expect(err).to.be.equals(error.message);
-          expect(barracks.getSegment).to.have.been.calledOnce;
-          expect(barracks.getSegment).to.have.been.calledWithExactly(token, changes.id);
-          expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-          expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('editSegment', {
-            headers: { 'x-auth-token': token },
-            body: Object.assign({}, originalSegment, changes),
-            pathVariables: { id: originalSegment.id }
-          });
-          done();
-        } catch (e) {
-          done(e);
-        }
-      });
-    });
-
-    it('should return an error message when the segment does not exist', done => {
-      // Given
-      const changes = { id: originalSegment.id, name: 'test segment' };
-      const error = { message: 'Error !' };
-      const excpectedResult = Object.assign({}, originalSegment, changes);
-      barracks.getSegment = sinon.stub().returns(Promise.reject(error));
-      barracks.client.sendEndpointRequest = sinon.spy();
-
-      // When / Then
-      barracks.editSegment(token, changes).then(result => {
-        done('Should have failed');
-      }).catch(err => {
-        try {
-          expect(err).to.be.equals(error.message);
-          expect(barracks.getSegment).to.have.been.calledOnce;
-          expect(barracks.getSegment).to.have.been.calledWithExactly(token, changes.id);
-          expect(barracks.client.sendEndpointRequest).to.not.have.been.called;
-          done();
-        } catch (e) {
-          done(e);
-        }
-      });
-    });
-
-    it('should return the edited segment', done => {
-      // Given
-      const changes = { id: originalSegment.id, name: 'NaaaAAmE' };
-      const expectedResult = Object.assign({}, originalSegment, changes);
-      const response = { body: expectedResult };
-      barracks.getSegment = sinon.stub().returns(Promise.resolve(originalSegment));
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.editSegment(token, changes).then(result => {
-        expect(result).to.be.equal(expectedResult);
-        expect(barracks.getSegment).to.have.been.calledOnce;
-        expect(barracks.getSegment).to.have.been.calledWithExactly(token, changes.id);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('editSegment', {
-          headers: { 'x-auth-token': token },
-          body: expectedResult,
-          pathVariables: { id: originalSegment.id }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-  });
-
-  describe('#getSegment()', () => {
-
-    const existingSegment = {
-      id: '123456789',
-      name: 'Plop',
-      query: {}
-    };
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getSegment(token, existingSegment.id).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegment', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            id: existingSegment.id
-          }
-        });
-        done();
-      });
-    });
-
-    it('should return a segment when request succeed', done => {
-      // Given
-      const response = { body: existingSegment };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegment(token, existingSegment.id).then(result => {
-        expect(result).to.be.equals(existingSegment);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegment', {
-          headers: { 'x-auth-token': token },
-          pathVariables: {
-            id: existingSegment.id
-          }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
     });
   });
 
@@ -310,259 +178,6 @@ describe('Barracks', () => {
         done();
       });
     });
-  });
-
-  describe('#getSegmentByName()', () => {
-
-    const otherSegment = { id: 'other', name: 'Other'};
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const segmentName = 'segment prod';
-      const error = 'Error !';
-      barracks.getSegments = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error);
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      });
-    });
-
-    it('should return an error if segment does not exist', done => {
-      // Given
-      const segmentName = 'segment prod';
-      const response = {
-        active: [
-          { id: 'zxcvbnm', name: 'segment' },
-          { id: 'zxcvbnm', name: 'other segment' }
-        ],
-        inactive: [],
-        other: otherSegment
-      };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals('No matching segment.');
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      });
-    });
-
-    it('should return specified active segment info when request succeed', done => {
-      // Given
-      const segmentName = 'segment prod';
-      const segment = { id: 'lkjhgfdsa', name: segmentName };
-      const response = {
-        active: [ segment, { id: 'zxcvbnm', name: 'other segment' } ],
-        other: otherSegment,
-        inactive: []
-      };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        expect(result).to.be.equals(segment);
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return other segment when name is other when request succeed', done => {
-      // Given
-      const response = { active: [ { id: 'zxcvbnm', name: 'other segment' } ], inactive: [], other: otherSegment };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, otherSegment.name).then(result => {
-        expect(result).to.be.equals(otherSegment);
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return specified segment when inactive', done => {
-      // Given
-      const segmentName = 'segment';
-      const segment = { id: 'lkjhgfdsa', name: segmentName };
-      const response = {
-        inactive: [ segment, { id: 'zxcvbnm', name: 'other segment' } ],
-        active: [],
-        other: otherSegment
-      };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      }).catch(err => {
-        done('err');
-      });
-    });
-
-    it('should return specified segment when inactive and active segments exist', done => {
-      // Given
-      const segmentName = 'segment prod';
-      const segment = { id: 'lkjhgfdsa', name: segmentName };
-      const response = {
-        inactive: [ segment, { id: 'zxcvbnm', name: 'other segment' } ],
-        active: [ { id: 'azdqshsd', name: 'wednesday segment' } ],
-        other: otherSegment
-      };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should reject an error when no inactive or active segment', done => {
-      // Given
-      const response = {
-        inactive: [],
-        active: [],
-        other: otherSegment
-      };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, 'segmentname').then(result => {
-        done('Should not have worked');
-      }).catch(err => {
-        expect(err).to.be.equal('No matching segment.')
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      });
-    });
-
-    it('should return specified segment when no inactive segment', done => {
-      // Given
-      const segmentName = 'segment prod';
-      const segment = { id: 'lkjhgfdsa', name: segmentName };
-      const response = { inactive: [], active: [ segment, { id: 'zxcvbnm', name: 'other segment' } ] };
-      barracks.getSegments = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegmentByName(token, segmentName).then(result => {
-        expect(barracks.getSegments).to.have.been.calledOnce;
-        expect(barracks.getSegments).to.have.been.calledWithExactly(token);
-        done();
-      }).catch(err => {
-        done(err)
-      });
-    });
-  });
-
-  describe('#getSegments()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.getSegments(token).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegments', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      });
-    });
-
-    it('should return segment list when request succeed', done => {
-      // Given
-      const segment1 = { id: 'lkjhgfdsa', name: 'prod' };
-      const segment2 = { id: 'qwertyuio', name: 'alpha' };
-      const segment3 = { id: 'vyugyu', name: 'plop' };
-      const activeSegments = [ segment1, segment2 ];
-      const inactiveSegments = [ segment3 ];
-      const body = { active: activeSegments, inactive: inactiveSegments };
-      const response = { body };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegments(token).then(result => {
-        expect(result).to.be.equals(body);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegments', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#createSegment()', () => {
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const segment = { name: 'Segment', query: { eq: { unitId: 'value' } } };
-      const error = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(error));
-
-      // When / Then
-      barracks.createSegment(token, segment).then(result => {
-        done('should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(error.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createSegment', {
-          headers: { 'x-auth-token': token },
-          body: segment
-        });
-        done();
-      });
-    });
-
-    it('should return the created segment', done => {
-      // Given
-      const segment = { name: 'Segment', query: { eq: { unitId: 'value' } } };
-      const savedSegment = Object.assign({}, segment, { userId: '123456789' });
-      const response = { body: savedSegment };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.createSegment(token, segment).then(result => {
-        expect(result).to.be.equals(savedSegment);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('createSegment', {
-          headers: { 'x-auth-token': token },
-          body: segment
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
   });
 
   describe('#createFilter()', () => {
@@ -1068,94 +683,6 @@ describe('Barracks', () => {
           'events',
           sinon.match.func
         );
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-  });
-
-  describe('#getSegments()', () => {
-
-    it('should return the user segments', done => {
-      // Given
-      const segments = {
-        active: [ buildSegment(1), buildSegment(2) ],
-        inactive: [ buildSegment(3), buildSegment(4) ]
-      };
-      const response = { body: segments };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.getSegments(token).then(result => {
-        expect(result).to.be.equals(segments);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegments', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      }).catch(err => {
-        done(err);
-      });
-    });
-
-    it('should return an error when the request failed', done => {
-      // Given
-      const errorMessage = "Error";
-      const errorResponse = { message: errorMessage };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(errorResponse));
-
-      // When / Then
-      barracks.getSegments(token).then(result => {
-        done('Should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(errorMessage);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('getSegments', {
-          headers: { 'x-auth-token': token }
-        });
-        done();
-      });
-    });
-
-  });
-
-  describe('#setActiveSegments()', () => {
-
-    const segmentIds = [ '12345', '67890' ];
-
-    it('should return an error message when request fails', done => {
-      // Given
-      const errorResponse = { message: 'Error !' };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.reject(errorResponse));
-
-      // When / Then
-      barracks.setActiveSegments(token, segmentIds).then(result => {
-        done('Should have failed');
-      }).catch(err => {
-        expect(err).to.be.equals(errorResponse.message);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('setActiveSegments', {
-          headers: { 'x-auth-token': token },
-          body: segmentIds
-        });
-        done();
-      });
-    });
-
-    it('should return segment ids when request is successful', done => {
-      // Given
-      const response = { body: segmentIds };
-      barracks.client.sendEndpointRequest = sinon.stub().returns(Promise.resolve(response));
-
-      // When / Then
-      barracks.setActiveSegments(token, segmentIds).then(result => {
-        expect(result).to.be.equals(segmentIds);
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledOnce;
-        expect(barracks.client.sendEndpointRequest).to.have.been.calledWithExactly('setActiveSegments', {
-          headers: { 'x-auth-token': token },
-          body: segmentIds
-        });
         done();
       }).catch(err => {
         done(err);
